Move PropertyFilter implementation to property_filter.rs
