+++ rust/feature-flags/src/flags/mod.rs
@@ -9,6 +9,7 @@ pub mod flag_operations;
 pub mod flag_property_group;
 pub mod flag_request;
 pub mod flag_service;
+pub mod property_filter;
 
 #[cfg(test)]
 mod test_flag_matching;
